Make OutboundDeliveryContext not anymore an inner class of EventBus
